// sveltekit/src/lib/proto/generated/global_settings.ts
export enum GlobalSettings_PipStyle {
    /**
     * @generated from protobuf enum value: Unknown = 0;
     */
    Unknown = 0,
    /**
     * @generated from protobuf enum value: None = 1;
     */
    None = 1,
    /**
     * @generated from protobuf enum value: Number = 2;
     */
    Number = 2,
    /**
     * @generated from protobuf enum value: Pips = 3;
     */
    Pips = 3,
    /**
     * @generated from protobuf enum value: PipsReal = 4;
     */
    PipsReal = 4
}